S-14b extract docs and research viewers to web/modules/docs-research.js
Move loadDocs, loadResearch, renderMarkdown, mdInline, escapeAttr and
downloadMarkdown from app.js into docs-research.js. Both pages are read-only
viewers sharing the same Markdown renderer with no cross-page state. Retain
local escapeAttr in app.js for the approvals section that remains there.
Fix ASI hazard: add explicit semicolon after boot() call, which was followed
directly by a (function ...) IIFE after the docs section was removed. Update
md-rendering-unify.test.ts to point assertions at docs-research.js.
